django/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
boto3==1.40.21
botocore==1.40.21
brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
commonmark==0.9.1
coverage==7.10.6
cryptography==45.0.6
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
django==5.2.5
django-allauth==65.11.1
django-cors-headers==4.7.0
django-extensions==4.1
django-templated-mail==1.1.1
django-unixdatetimefield==1.2.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
eyed3==0.9.8
filetype==1.2.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.30.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
moviepy==2.2.1
mutagen==1.47.0
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.32.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==12.6.0
rsa==4.9.1
s3transfer==0.13.1
sentry-sdk==1.45.1
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
spotify-dl==8.9.0
spotipy==2.25.1
sqlparse==0.5.3
toml==0.10.2
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
websockets==15.0.1
youtube-dl==2021.12.17
yt-dlp==2025.8.27